requirements.txt
streamlit
pandas
fpdf2
google-generativeai
PyMuPDF
